+++ packages/acquirer-frontend/src/components/ui/FileUploadModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FileUploadModal'
